requirements.txt
PyGithub==2.5.0
python-dotenv==1.0.1
aiogram==3.17.0
